Update View를 나누어 적용
